Code cleanup - move ItineraryMapper to package org.opentripplanner.routing.algorithm.mapping.
