+++ src/main/java/org/opentripplanner/routing/algorithm/mapping/ItineraryMapper.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.algorithm.raptor.itinerary;
+package org.opentripplanner.routing.algorithm.mapping;
 
 import org.locationtech.jts.geom.Coordinate;
 import org.opentripplanner.api.model.Itinerary;
@@ -23,7 +23,6 @@ import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.error.TrivialPathException;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Vertex;
-import org.opentripplanner.routing.services.FareService;
 import org.opentripplanner.routing.spt.GraphPath;
 import org.opentripplanner.routing.trippattern.TripTimes;
 import org.opentripplanner.routing.vertextype.TransitStopVertex;
